# perf: improve initial css generation time by 5x