--- a/packages/core/__tests__/fixture.ts
+++ b/packages/core/__tests__/fixture.ts
@@ -1,13 +1,11 @@
 import * as mocks from '@pandacss/fixture'
 import { TokenDictionary } from '@pandacss/token-dictionary'
-import postcss from 'postcss'
-import { Conditions, Utility, Recipes } from '../src'
+import { Dict } from '@pandacss/types'
+import { Conditions, Recipes, Utility, createSheetRoot } from '../src'
+import { createAtomicRule } from '../src/atomic-rule'
 import type { StylesheetContext } from '../src/types'
 
-type ContextOptions = {
-  hash?: boolean
-  prefix?: string
-}
+type ContextOptions = Partial<Omit<StylesheetContext, 'recipes'>> & { prefix?: string; recipes?: Dict }
 
 export const defaultLayers = {
   reset: 'reset',
@@ -17,7 +15,9 @@ export const defaultLayers = {
   utilities: 'utilities',
 }
 
-export const createContext = ({ hash, prefix }: ContextOptions = {}): StylesheetContext => {
+export const createContext = (opts: ContextOptions = {}): StylesheetContext => {
+  const { hash, prefix, recipes: recipeObj = {}, ...rest } = opts
+
   const conditions = new Conditions({
     conditions: mocks.conditions,
     breakpoints: mocks.breakpoints,
@@ -36,35 +36,64 @@ export const createContext = ({ hash, prefix }: ContextOptions = {}): Stylesheet
     shorthands: true,
   })
 
+  const sheetRoot = createSheetRoot(defaultLayers)
+
+  const recipes = new Recipes(recipeObj, {
+    utility,
+    conditions,
+    layers: sheetRoot.layers,
+  })
+
   return {
+    ...sheetRoot,
     hash,
-    root: postcss.root(),
-    conditions: conditions,
-    utility: utility,
+    recipes,
+    conditions,
+    utility,
     helpers: { map: () => '' },
-    layers: defaultLayers,
+    ...rest,
   }
 }
 
+export const createCssFn =
+  (opts: ContextOptions = {}) =>
+  (styles: Dict) => {
+    const ctx = createContext(opts)
+    const rule = createAtomicRule(ctx)
+    rule.process({ styles })
+    ctx.insertLayers()
+    return ctx.root.toString()
+  }
+
+export const createRecipeFn =
+  (opts: ContextOptions = {}) =>
+  (recipe: string, styles: Dict) => {
+    const ctx = createContext(opts)
+    ctx.recipes.process(recipe, { styles })
+    ctx.insertLayers()
+    return ctx.root.toString()
+  }
+
 export function getRecipe(key: 'buttonStyle' | 'textStyle' | 'tooltipStyle') {
-  const recipes = new Recipes(mocks.recipes, createContext())
-  recipes.save()
-  const recipe = recipes.getRecipe(key)
-  return recipe!.config
+  const ctx = createContext()
+  const recipes = new Recipes(mocks.recipes, ctx)
+  return recipes.getRecipe(key)!
 }
 
-export function processRecipe(recipe: 'buttonStyle' | 'textStyle' | 'tooltipStyle', value: Record<string, any>) {
-  const recipes = new Recipes(mocks.recipes, createContext())
-  recipes.save()
-  recipes.process(recipe, { styles: value })
-  return recipes.toCss()
+export function getSlotRecipe(key: 'button') {
+  const ctx = createContext()
+  const recipes = new Recipes(mocks.slotRecipes, ctx)
+  return recipes.getRecipe(key)!
 }
 
-export function processSlotRecipe(recipe: 'button', value: Record<string, any>) {
-  const recipes = new Recipes(mocks.slotRecipes, createContext())
-  recipes.save()
-  recipes.process(recipe, { styles: value })
-  return recipes
+export function processRecipe(recipe: 'buttonStyle' | 'textStyle' | 'tooltipStyle', styles: Record<string, any>) {
+  const recipeFn = createRecipeFn({ recipes: mocks.recipes })
+  return recipeFn(recipe, styles)
+}
+
+export function processSlotRecipe(recipe: 'button', styles: Record<string, any>) {
+  const recipeFn = createRecipeFn({ recipes: mocks.slotRecipes })
+  return recipeFn(recipe, styles)
 }
 
 export const compositions = {