--- a/packages/core/src/index.ts
+++ b/packages/core/src/index.ts
@@ -4,14 +4,22 @@ export { assignCompositions } from './compositions'
 export { Conditions } from './conditions'
 export { isSlotRecipe } from './is-slot-recipe'
 export { toKeyframeCss } from './keyframes'
+export { Layers } from './layers'
 export { mergeCss } from './merge-css'
 export { expandCssFunctions, expandNestedCss, optimizeCss, prettifyCss } from './optimize'
 export { Recipes } from './recipes'
 export { extractParentSelectors } from './selector'
 export { sharedHooks } from './shared-hooks'
-export { createSheetRoot } from './sheet-root'
 export { getStaticCss } from './static-css'
 export { Stylesheet } from './stylesheet'
 export { toCss } from './to-css'
-export type { RecipeNode, StylesheetContext, StylesheetLayers, StylesheetRoot, TransformResult } from './types'
+export type {
+  RecipeNode,
+  StylesheetContext,
+  StylesheetLayers,
+  StylesheetRoot,
+  TransformResult,
+  RecipeContext,
+  AtomicRuleContext,
+} from './types'
 export { Utility } from './utility'