--- a/packages/vite-plugin-studio/index.ts
+++ b/packages/vite-plugin-studio/index.ts
@@ -1,0 +1,1 @@
+export * from './src'