Closes #1745 

## 📝 Description

This PR boosts the perf of the CSS extraction time by at least 5x.

**Park UI**
- Before regression: 3500ms
- After regression: 7400ms
- This PR: 580ms

**Website**
- Before regression: 2900ms
- This PR:  208ms

## ⛳️ Current behavior (updates)

Initial extraction time can get slow when
- using static CSS with lots of recipes
- the parsed files gets pretty large

## 🚀 New behavior

Much faster!

## 💣 Is this a breaking change (Yes/No):

- The only breaking change I can recall is that hooks no longer return the css.

## 📝 Additional Information